Clean up typing in serial protocol and add post-mortem debug logging

+++ crates/ziggurat-esp/src/api.rs
@@ -798,7 +798,7 @@ fn notification_to_json(notification_event: ZigbeeNotification) -> Value {
                 "source": hex::encode(source.to_bytes()),
                 "source_ieee": eui64_to_string(source_ieee),
                 "frame_counter": frame_counter,
-                "key_id": key_id,
+                "key_id": key_id as u8,
             }),
         ),
         ZigbeeNotification::SendConfirm { request_id, result } => notification(
